% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARZIMM.control.R
\name{ARZIMM.control}
\alias{ARZIMM.control}
\title{internal ARZIMM parameters}
\usage{
ARZIMM.control(
  initpara = list(betaini = NULL, gammaini = NULL, sigmaini = NULL, weight1all = NULL,
    weight2all = NULL, calini = TRUE, iniw = FALSE),
  tunpara = list(lambdaseq1 = NULL, lambdaseq2 = NULL, ntun = 50, epsilon = c(50000, 5)),
  selectpara = list(selgamma = TRUE, selcri = "BIC", pen = "adalasso"),
  bootpara = list(bootpval = TRUE, nboot = 500)
)
}
\arguments{
\item{initpara}{a named list of user-specified initial values:
\describe{
  \item{betaini}{the matrix of fixed effects for the non-zero auto-regressive model}
  \item{gammaini}{the matrix of fixed effects for the zero state logit model}
  \item{sigmaini}{the measurement error standard deviation for both the non-zero auto-regressive model and
  the zero state logit model}
  \item{weight1all}{observation weights for the non-zero auto-regressive model. Default is 1 for each observation}
  \item{weight2all}{observation weights for the zero state logit model. Default is 1 for each observation}
  \item{calini=T}{logical; should initial parameters be calculated. Default is Ture}
  \item{iniw}{logical; should observation weights be calculated according to initial parameters. Default is False}
  When this list of initial values does not contain some of these components or contains components
  not of the appropriate length, then the default initial values are used instead.
}}

\item{tunpara}{a list of control values with components:
\describe{
  \item{lambdaseq1}{a user supplied \code{lambda} sequence for the non-zero auto-regressive model.
  Typical usage is to have the program compute its own \code{lambda} sequence. Supplying a value of
  \code{lambda} overrides this. WARNING: use with care. Avoid supplying a single value for \code{lambda}}
  \item{lambdaseq2}{a user supplied \code{lambda} sequence for the  zero state logit model.
  Typical usage is to have the program compute its own \code{lambda} sequence. Supplying a value of
  \code{lambda} overrides this. WARNING: use with care. Avoid supplying a single value for \code{lambda}}
  \item{ntun}{the number of \code{lambda} values - default is 50}
  \item{epsilon}{the range of lambda values; default= c(5e4,5)}
}}

\item{selectpara}{a list of control values with components:
\describe{
  \item{selgamma}{logical; should concomitant variables in the zero state logit model be selected}
  \item{selcri}{method to be use for variable selection. Currently three options. The default is BIC.
  Other choices are AIC and CV}
  \item{pen}{penalty; defualt is 'adalasso'}
}}

\item{bootpara}{a list of control values with components:
\describe{
  \item{bootpval}{logical; should p value be calculated using bootstrap}
  \item{nboot}{the number of bootstrap simulations; default is 500}
}}
}
\value{
A list with named elements as in the argument list
}
\description{
View and/or change the factory default parameters in ARZIMM
}
\details{
If called with no arguments, \code{ARZIMM.control()} returns a list with the
current settings of these parameters. Any arguments included in the call
sets those parameters to the new values, and then silently returns. The
values set are persistent for the duration of the R session.
}
\examples{

ARZIMM.control(bootpara=list(bootpval=FALSE))
ARZIMM.control(selectpara=list(selgamma=FALSE,pen='lasso'))

}
\seealso{
\code{ARZIMM}
}
\author{
Linchen He \cr Maintainer: Linchen He
\email{Linchen.He@nyulangone.org}
}
\keyword{controls}
\keyword{model}
