% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARZIMM.R
\name{ARZIMM}
\alias{ARZIMM}
\title{Auto-Regressive Zero-Inflated Mixed Model}
\usage{
ARZIMM(
  Varname = NA,
  Conname = NA,
  Covname = NA,
  fdata = NA,
  IDname = "ID",
  Tname = "Time",
  group = NA,
  family = "Poisson",
  initpara = list(betaini = NULL, gammaini = NULL, sigmaini = NULL, weight1all = NULL,
    weight2all = NULL, calini = T, iniw = F),
  tunpara = list(lambdaseq1 = NULL, lambdaseq2 = NULL, ntun = 50, epsilon = c(50000, 5)),
  selectpara = list(selgamma = T, selcri = "BIC", pen = "adalasso"),
  bootpara = list(bootpval = T, nboot = 500)
)
}
\arguments{
\item{Varname}{a vector of character string indicating the taxa names in the non-zero auto-regressive model}

\item{Conname}{a vector of character string indicating the concomitant variable names in the zero state logit model}

\item{Covname}{a vector of character string indicating the covariate names in the non-zero auto-regressive model}

\item{fdata}{a data frame containing all variables to be analysized}

\item{IDname}{a character string indicating the subject ID. Default is ID}

\item{Tname}{a character string indicating the time variable. Default is Time}

\item{group}{a vector of numbers as group indicator. To be calculated automatically When absent}

\item{family}{a character string indicating the distribtuion. default is Poisson}

\item{initpara}{a named list of user-specified initial values:
\describe{
  \item{betaini}{the matrix of fixed effects for the non-zero auto-regressive model}
  \item{gammaini}{the matrix of fixed effects for the zero state logit model}
  \item{sigmaini}{the measurement error standard deviation for both the non-zero auto-regressive model and
  the zero state logit model}
  \item{weight1all}{observation weights for the non-zero auto-regressive model. Default is 1 for each observation}
  \item{weight2all}{observation weights for the zero state logit model. Default is 1 for each observation}
  \item{calini=T}{logical; should initial parameters be calculated. Default is Ture}
  \item{iniw}{logical; should observation weights be calculated according to initial parameters. Default is False}
  When this list of initial values does not contain some of these components or contains components
  not of the appropriate length, then the default initial values are used instead.
}}

\item{tunpara}{a list of control values with components:
\describe{
  \item{lambdaseq1}{a user supplied \code{lambda} sequence for the non-zero auto-regressive model.
  Typical usage is to have the program compute its own \code{lambda} sequence. Supplying a value of
  \code{lambda} overrides this. WARNING: use with care. Avoid supplying a single value for \code{lambda}}
  \item{lambdaseq2}{a user supplied \code{lambda} sequence for the  zero state logit model.
  Typical usage is to have the program compute its own \code{lambda} sequence. Supplying a value of
  \code{lambda} overrides this. WARNING: use with care. Avoid supplying a single value for \code{lambda}}
  \item{ntun}{the number of \code{lambda} values - default is 50}
  \item{epsilon}{the range of lambda values; default= c(5e4,5)}
}}

\item{selectpara}{a list of control values with components:
\describe{
  \item{selgamma}{logical; should concomitant variables in the zero state logit model be selected}
  \item{selcri}{method to be use for variable selection. Currently three options. The default is BIC.
  Other choices are AIC and CV}
  \item{pen}{penalty; defualt is 'adalasso'}
}}

\item{bootpara}{a list of control values with components:
\describe{
  \item{bootpval}{logical; should p value be calculated using bootstrap}
  \item{nboot}{the number of bootstrap simulations; default is 500}
}}
}
\value{
an object of class \code{"ARZIMMObject"} is returned, which is a list with the ingredients of fit.
  \item{nwtable}{the matrix of network table of fit}
  \item{mseest}{a list of mean square error:
  \describe{
    \item{rmseest}{the root square of mean standard error}
    \item{rmseest}{the root square of mean pearson standardized error}
  }}
  \item{paralist}{a list of parameter estimates:
  \describe{
    \item{beta}{the matrix of fixed effects for the non-zero auto-regressive model}
    \item{gamma}{the matrix of fixed effects for the zero state logit model}
    \item{sigma}{the measurement error standard deviation for both the non-zero auto-regressive model and
    the zero state logit model}
    \item{ciest}{the estimated random effects part of both the non-zero auto-regressive model and
    the zero state logit model}
  }}
  \item{runtime}{running time of the program}
  \item{datalist}{an object of class \code{"ARZIMMData"}}
  \item{resultall}{a list of parameter estimates of the fits with ingredients of the lambda}
  \item{bootparapval}{a list of p values obtain via bootstrap with componenets:
  \describe{
    \item{betapval}{a vector of p values of fixed effects for the non-zero auto-regressive model}
    \item{gammapval}{a vector of p values of fixed effects for the zero state logit model}
    \item{sigmapval}{a vector of p values of the measurement error standard deviation for both the
    non-zero auto-regressive model and the zero state logit model}
  }}
  \item{tunlist}{the values of parameters used in the fits with components:
  \describe{
    \item{lambdabeta}{the values of \code{lambda} used in the non-zero auto-regressive model}
    \item{lambdagamma}{the values of \code{lambda} used in the zero state logit model}
    \item{weight1all}{observation weights used in the non-zero auto-regressive model}
  }}
  \item{parasetup}{a list of parameters used to initial the ARZIMM program with components:
  \describe{
    \item{initpara}{a list of initial parameter inputs; if the inputs are absent, default values are included}
    \item{lambda}{the values of \code{lambda} used in the fits.}
    \item{tunpara}{a list of tunning parameter inputs; if the inputs are absent, default values are included}
    \item{selectpara}{a list of selection parameter inputs; if the inputs are absent, default values are included}
  }}
}
\description{
This function allows you to fit the ARZIMM.
}
\examples{

data(sampledata)
ARZIMM::ARZIMM(Varname = Varname,Conname = Conname,fdata = fdata,IDname = IDname,Tname = Tname,
bootpara=list(bootpval=T,nboot=100))
}
\keyword{ARZIMM}
