% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emnr.R
\name{emnr}
\alias{emnr}
\title{EM algorithm}
\usage{
emnr(data, para, weight, family, selgamma = F)
}
\arguments{
\item{data}{a list of data with componenets:
\describe{
  \item{yFdata}{a vector of}
  \item{xFdata}{a matrix of variables corresponding to the non-zero auto-regressive model}
  \item{conFdata}{a matrix of covariates corresponding to the zero state logit model}
  \item{group}{a vector of numbers as group indicator}
}}

\item{para}{a list of parameter estimates:
\describe{
  \item{beta}{initial value for beta}
  \item{gamma}{initial value for gamma}
  \item{sigma}{initial value for sigma}
}}

\item{weight}{a vector of observation weightsfor both the non-zero auto-regressive model and
the zero state logit model}

\item{family}{a character string indicating the distribtuion. default is Poisson}

\item{selgamma}{logical; should concomitant variables in the zero state logit model be selected}
}
\value{
a list of fits
  \item{para}{a list of parameter estimates:
  \describe{
    \item{beta}{beta estimates}
    \item{gamma}{gamma estimates}
    \item{sigma}{sigma estimates}
  }}
  \item{ciestm}{the estimated random effects}
  \item{conv}{logical; did the algorithm converged}
  \item{df}{the number of non-zero parameter estimates}
  \item{bic}{a vector of BIC, AIC, and log likelihood}
  \item{mse}{a vector of square root of mean (pearson) standard error}
  \item{lambda}{a vector of \code{lambda} sequence}
}
\description{
EM algorithm used in the ARZIMM model.
}
\examples{
emnr()
}
\keyword{EM}
\keyword{algorithm}
