% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMixTime.R
\name{simMixTime}
\alias{simMixTime}
\title{A simulation function}
\usage{
simMixTime(
  baseFdataN2M,
  conFdataN2L,
  timeN,
  interceptM,
  betaMM,
  gammaLM = NULL,
  sigmaM,
  biN2M = NULL,
  family = "Poisson"
)
}
\arguments{
\item{baseFdataN2M}{a matrix of absolute counts at baseline time}

\item{conFdataN2L}{a matrix of concomitant variables}

\item{timeN}{a vector of the number of time points for each subjects}

\item{interceptM}{a vector of the intercepts of the non-zero autoregression model}

\item{betaMM}{a matrix of network table}

\item{gammaLM}{a matrix of the zero state logit model}

\item{sigmaM}{a vector of the measurement error standard deviation for the non-zero autoregression model}

\item{biN2M}{a vector of the random effects for the non-zero autoregression model}

\item{family}{a character string indicating the distribtuion. default is Poisson}
}
\value{
a list of simulations:
  \item{otu.tab}{a data frame of OTU table}
  \item{convar}{a data frame of concomitant variables}
  \item{subjid}{a vector of subject-time IDs}
  \item{df}{the number of non-zero parameter estimates}
  \item{zi}{a list of the random effects}
}
\description{
Simulation for the ARZIMM model.
}
\examples{
simMixTime()
}
\keyword{Simulation}
